+++ ThirdPartyAdapters/chartboost/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Chartboost Android Mediation Adapter Changelog
 
+#### Next Version
+- Updated `minSdk` to 23
+- Rewarded ads now calls `onUserEarnedReward` without parameters.
+
 #### Version 9.8.3.0
 - Verified compatibility with Chartboost SDK version 9.8.3.